cli/pyproject.toml
[tool.poetry]
name = "nexus-cli"
version = "1.0.0"
description = "Command-line interface for sales tax nexus analysis"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "nexus_cli", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.7"
pandas = "^2.0.0"
pyyaml = "^6.0.1"
pydantic = "^2.0.0"
rich = "^13.7.0"
openpyxl = "^3.1.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.12.0"
mypy = "^1.7.0"
ruff = "^0.1.8"

[tool.poetry.scripts]
nexus = "nexus_cli.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=nexus_cli --cov-report=term-missing"
